+++ pybricks/util_mp/pb_obj_helper.h
@@ -30,6 +30,9 @@ mp_obj_t pb_obj_new_fraction(int32_t numerator, int32_t denominator);
 // Get absolute value if object is not none, else return default
 mp_int_t pb_obj_get_default_abs_int(mp_obj_t obj, mp_int_t default_val);
 
+// Get array of percentages from single value or tuple
+void pb_obj_get_pct_or_array(mp_obj_t obj_in, size_t num, int8_t *values);
+
 // Get base instance if object is instance of subclass of type
 mp_obj_t pb_obj_get_base_class_obj(mp_obj_t obj, const mp_obj_type_t *type);